pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ddns-update-monkey"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "requests~=2.32",
    "python-dotenv~=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/monkey"]
